A little more cleanup of the Cache
